Move Agent installer from common to datadog/agent
