+++ datadog/agent/install.go
@@ -1,4 +1,4 @@
-package agentinstall
+package agent
 
 import (
 	"strings"
